single sdpa function
