+++ llms/mlx_lm/generate.py
@@ -7,7 +7,7 @@ import sys
 import mlx.core as mx
 
 from .models.cache import QuantizedKVCache, load_prompt_cache
-from .utils import check_quantized_kv_args, generate, load
+from .utils import generate, load
 
 DEFAULT_PROMPT = "hello"
 DEFAULT_MAX_TOKENS = 100
@@ -15,6 +15,7 @@ DEFAULT_TEMP = 0.0
 DEFAULT_TOP_P = 1.0
 DEFAULT_SEED = 0
 DEFAULT_MODEL = "mlx-community/Llama-3.2-3B-Instruct-4bit"
+DEFAULT_QUANTIZED_KV_START = 5000
 
 
 def str2bool(string):
@@ -108,24 +109,24 @@ def setup_arg_parser():
         help="A file containing saved KV caches to avoid recomputing them",
     )
     parser.add_argument(
-        "--quantized-kv-start",
-        help="Use a quantized KV cache from this step onwards.",
+        "--kv-bits",
         type=int,
+        help="Number of bits for KV cache quantization. "
+        "Defaults to no quantization.",
         default=None,
     )
     parser.add_argument(
         "--kv-group-size",
         type=int,
-        help="Group size for kv cache quantization. "
-        "--quantized-kv-start must be provided to have an effect.",
+        help="Group size for KV cache quantization.",
         default=64,
     )
     parser.add_argument(
-        "--kv-bits",
+        "--quantized-kv-start",
+        help="When --kv-bits is set, start quantizing the KV cache "
+        "from this step onwards.",
         type=int,
-        help="Number of bits for kv cache quantization. "
-        "--quantized-kv-start must be provided to have an effect.",
-        default=8,
+        default=DEFAULT_QUANTIZED_KV_START,
     )
     return parser
 
@@ -173,13 +174,15 @@ def main():
             args.prompt_cache_file,
             return_metadata=True,
         )
-        if args.quantized_kv_start and isinstance(prompt_cache[0], QuantizedKVCache):
-            raise ValueError(
-                "Specified `--quantized-kv-start` but cache from "
-                "`--prompt-cache-file` is already quantized."
-            )
-
-    check_quantized_kv_args(args.quantized_kv_start, args.kv_group_size, args.kv_bits)
+        if isinstance(prompt_cache[0], QuantizedKVCache):
+            if args.kv_bits is not None and args.kv_bits != prompt_cache[0].bits:
+                raise ValueError(
+                    "--kv-bits does not match the kv cache loaded from --prompt-cache-file."
+                )
+            if args.kv_group_size != prompt_cache[0].group_size:
+                raise ValueError(
+                    "--kv-group-size does not match the kv cache loaded from --prompt-cache-file."
+                )
 
     # Building tokenizer_config
     tokenizer_config = (
