+++ llms/mlx_lm/models/phixtral.py
@@ -8,7 +8,7 @@ from typing import Tuple
 import mlx.core as mx
 import mlx.nn as nn
 
-from .base import create_attention_mask
+from .base import create_attention_mask, scaled_dot_product_attention
 from .switch_layers import SwitchMLP
 
 
@@ -71,7 +71,7 @@ class RoPEAttention(nn.Module):
         # Finally perform the attention computation
         scale = math.sqrt(1 / queries.shape[-1])
 
-        output = mx.fast.scaled_dot_product_attention(
+        output = scaled_dot_product_attention(
             queries.astype(mx.float32), keys, values, scale=scale, mask=mask
         ).astype(values.dtype)
         output = output.moveaxis(2, 1).reshape(B, L, -1)
